LanguageAppBackend/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-speech==2.12.0
googleapis-common-protos==1.54.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
libcst==0.4.1
MarkupSafe==2.1.0
mypy-extensions==0.4.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
